+++ testing/tiny.c
@@ -1,0 +1,10 @@
+#include "../littleHelpers.h"
+
+// #define ERR_MULTIPROCESS 0
+// #define USAGE_STRING ""
+#include "../easyCheck.h"
+
+int main(int argc, char** argv) {
+    UNUSED(argc);
+    UNUSED(argv);
+}
